UUIDs

+++ app/services/spree_avatax_official/create_tax_adjustments.rb
@@ -1,15 +1,15 @@
 module SpreeAvataxOfficial
   class CreateTaxAdjustments < SpreeAvataxOfficial::Base
     def call(order:)
+      return failure(I18n.t('spree_avatax_official.create_tax_adjustments.order_canceled')) if order.canceled?
+
       order.reload.all_adjustments.tax.destroy_all
 
       avatax_response = send_transaction_to_avatax(order)
 
       return failure(false) if avatax_failed_response?(avatax_response)
 
-      avatax_response.value['lines'].each do |avatax_item|
-        process_avatax_item(order, avatax_item)
-      end
+      process_avatax_items(order, avatax_response.value['lines'])
 
       success(true)
     end
@@ -32,24 +32,29 @@ module SpreeAvataxOfficial
       avatax_response.failure? || avatax_response.value['totalTax'].zero?
     end
 
+    def process_avatax_items(order, avatax_items)
+      avatax_items.each { |avatax_item| process_avatax_item(order, avatax_item) }
+    end
+
     def process_avatax_item(order, avatax_item)
       tax_amount = avatax_item['taxCalculated']
 
       return if tax_amount.zero?
 
-      item_id, item_suffix = avatax_item['lineNumber'].split('-')
-      item                 = find_item(order, item_id, item_suffix)
-      tax_rate             = find_or_create_tax_rate(item)
+      item_suffix = avatax_item['lineNumber'].slice!(0..2)
+      item_id     = avatax_item['lineNumber']
+      item        = find_item(order, item_id, item_suffix)
+      tax_rate    = find_or_create_tax_rate(item)
 
       create_tax_adjustment(item, tax_rate, tax_amount)
     end
 
-    def find_item(order, id, suffix)
+    def find_item(order, uuid, suffix)
       case suffix
-      when 'LI'
-        order.line_items.find(id)
-      when 'FR'
-        order.shipments.find(id)
+      when 'LI-'
+        order.line_items.find_by(avatax_uuid: uuid)
+      when 'FR-'
+        order.shipments.find_by(avatax_uuid: uuid)
       end
     end
